+++ CHANGELOG.md
@@ -24,7 +24,8 @@ This changelog includes only user-sided updates. It indicates the advancement fo
 - ✔️ Cells animation
 - ✔ Interface animation
 - ✔ Last score information animation
-- ✔ Settings to toggle animation, game colors, and particles with a previewer
+- ✔ Settings to toggle animation, game colors, and particles
+- ✔ Settings previewer with a grid view
 - ✨ Changed score calculation to get an exponential one
 - 🏳 Statistics
 - 🏳 Sounds